refactor tests a bit more, include built file
